Make licence agreements import configurable (#367)
https://eaflood.atlassian.net/browse/WATER-3201

From our investigations, it looks like the import of licence agreements was a 'one-off' process intended to be run just the once when charging went live in WRLS. We have a flag in the config that controls whether they get included when the main licence import takes place.

That is currently hard coded to `false`. The commit history shows it was set to `true`, code was released and shipped, then immediately set to `false`. So, being pedantic, it ain't really config!!

It also turns out that any new developer on the team needs to be able to import the original NALD charge agreements when setting up their local environment. The same is true for any environment where we drop the DB and start again.

We need a way to enable this part of the import that does not require amending the code. This change adds a new env var that if present, will allow us to control whether licence agreements are imported or not.