Make cron background job timings configurable (#398)
* Make cron background job timings configurable

https://eaflood.atlassian.net/browse/WATER-3745

The context for this change is actually to give as much 'oomph' to our pre-prod environment to run some other testing. By making the background cron jobs in this service configurable we can ensure they run outside the time testing is taking place.

However, we would expect them to be configurable anyway to give us this kind of flexibility. So, this change amends the cron config in `config.js`. Now what is set for `production` will become the default if an env var with the timing is not set.

* Add examples to .env.example

* Update config.js to use new cron env vars

* Fix the tests

* removed unnessessary require of dotenv

Co-authored-by: Jason Claxton <30830544+Jozzey@users.noreply.github.com>