Update tests for core/plugin.js
NODE_ENV is used here to _actually_ check if we are in `test` mode. Having looked at https://github.com/DEFRA/water-abstraction-import/pull/292 which changed `if (process.env.NODE_ENV !== 'test') {` in `src/modules/core/plugin.js` away from checking if we were running on travis-ci, this seems to be a valid check.

Why this is the one place in all the jobs we trigger that cares if we are running in a true `test` environment we cannot explain! But we can leave the use of NODE_ENV for now.

We do the tweak the tests though. Clearly, they were thining in terms of environment. But by making `NODE_ENV=production` we should confirm that it's what `isProduction` returns that matters.