Correctly set NODE_ENV (#405)
https://github.com/DEFRA/water-abstraction-team/issues/43

We are currently using `NODE_ENV` incorrectly; we set it to be the name of the specific environment (eg. `local`, `qa` etc.) when it should be more like the _type_ of environment (eg. is it a dev environment, a production environment, etc.)

This means that tools like pm2 aren't correctly picking up that we are running in a dev-like or prod-like environment and configuring accordingly.

We, therefore, change how we use `NODE_ENV` to be the type of environment, with a separate `ENVIRONMENT` env var to set the actual environment (eg. `tst`, `pre` etc.)

** Notes

- Control etag check directly with env var
  Again, we have a complicated setup related to how env vars are controlling behaviour in the app. We clean up the config and give ourselves more flexibility by add a new env var to control the NALD etag check rather than using a proxy.

- Remove isLocal check
  This was just being used in the config for redis. The water-abstraction-service was built later and takes a different approach which we simply lift and apply here.

- Update tests for core/plugin.js and src/modules/nald-import/plugin.js
  NODE_ENV is used here to _actually_ check if we are in `test` mode. Having looked at https://github.com/DEFRA/water-abstraction-import/pull/292 which changed `if (process.env.NODE_ENV !== 'test') {` in `src/modules/core/plugin.js` away from checking if we were running on travis-ci, this seems to be a valid check. Why this is the one place in all the jobs we trigger that cares if we are running in a true `test` environment we cannot explain! But we can leave the use of NODE_ENV for now. We do the tweak the tests though. Clearly, they were thinking in terms of environment. But by making `NODE_ENV=production` we should confirm that it's what `isProduction` returns that matters.

- Fix plugin TRAVIS tests
  Clearly, we should be whipping out all references to travis-ci as per https://github.com/DEFRA/water-abstraction-team/issues/44. But for now we just need to fix the meaningless use of NODE_ENV in the tests. What it's set to has no impact on whether the subscribers get registered!

- Remove references in `test/modules/charging-import/plugin.js`
  It sets the NODE_ENV to various environments. However, the schedule for the charging import is now set directly via `WRLS_CRON_CHARGING`. So, whatever NODE_ENV is, is meaningless to what schedule is applied, especially in a test where we are stubbing the config value!

The final thing to note is we've changed the behaviour of the app slightly. Previously if any errors occurred in pre-prod or production it would send an email. We actually only care about errors in production. So, we also get the chance to simplify things by updating the code to just send an email for production errors.