Remove references to Travis-CI (#419)
https://github.com/DEFRA/water-abstraction-team/issues/44

There are references in the code to a `TRAVIS` env var. This must have been used when [Travis-CI](https://travis-ci.com) was used as the build service.

When they changed their pricing model some years ago all projects moved to GitHub Actions, including this one. But clearly, we didn't tidy up everything during the migration.

This change removes those last vestiges.

** Notes

Having made the change we realised the plugin was useless so we also updated the code to allow us to remove it.