Remove unit test file
The unit test file `index.test.js` was able to confirm that what we've done works; however as a test it isn't great simply because our server isn't set up for easy testing. At this stage we don't want to invest time in restructuring it just to test a single plugin so instead we remove the test file now that it's served its purpose.