Move `pg-boss` from helpers (#436)
* Move `pg-boss` in from helpers

`water-abstraction-helpers` exports functionality related to the `pg-boss` package. However `water-abstraction-import` is the only one of our services that actually uses this. We therefore move the functionality out of the helpers and into this service so that we're not importing `pg-boss` into all of our services needlessly.

* Simple test

* Move existing code into separate plugin file

We create a `src/plugins` folder and a file within that `pg-boss.plugin.js`. As an initial step, we simply move the existing plugin setup into that file and update `index.js` to use it.

* Add `pg-boss` dependency

We add the `pg-boss` dependency. We use the exact same version as in the helpers to ensure compatibility, as we know that later versions break our existing code so we will consider these changes at a later date.

* Transfer helper plugin to our code

We copy the plugin code in the helpers to our new `pg-boss.plugin` file

* Remove unit test file

The unit test file `index.test.js` was able to confirm that what we've done works; however as a test it isn't great simply because our server isn't set up for easy testing. At this stage we don't want to invest time in restructuring it just to test a single plugin so instead we remove the test file now that it's served its purpose.