Remove dead deep-map dependency
So, with our bump of the helpers everything broke! After investigation this was because `scripts/licence-creator/index.js` was requiring [deep-map](https://github.com/mcmath/deep-map) to do some work (see new comment in code to find out what that is).

The problem is **deep-map** was then requiring `es6-weak-map` but only had it listed as a dev dependency. This is raised [as an issue](https://github.com/mcmath/deep-map/issues/14) and someone has even gone to the effort of creating a [fix PR](https://github.com/mcmath/deep-map/pull/15). But to no avail as these were done in 2019 and multiple pings to the contributor have not resulted in the fix being merged.

The last release was 4 years ago so this is clearly an abandoned project. Also, we've been able to replicate what it was doing with very little extra code so it was unnecessary in the first place.