Fix issue with migrations (#502)
We have spotted that migrations are not working as expected. They failed when we deployed them to our AWS environment. When we investigated we found that a new `migrations` table had been created in the `public` schema. This meant it was believed no migrations had been run and so was trying to run them all.

This repo has 5 migrations. The first 4 can be repeatedly run because they either initiate commands where the current state of the DB doesn't matter, or they have conditions in place, for example, `CREATE TABLE IF NOT EXISTS`. But the 5th uses a simple `CREATE TABLE water_import.charge_versions_metadata` which fails because the table already exists.

The root cause is our mistaken assumption that providing the `DATABASE_URL` env var means [db-migrate](https://db-migrate.readthedocs.io/) doesn't use `database.json`. In fact, it continues to use the 2 together; `DATABASE_URL` just takes precedence.

The schema isn't specified in the URL so **db-migrate** was falling back to `database.json`. When we removed it in <https://github.com/DEFRA/water-abstraction-import/pull/493> it meant **db-migrate** was falling all the way back to using the `public` schema.

So, this change adds the file back in but just specifies the schema to use.

** Notes

If using a config file **db-migrate** requires that it contains a `dev:` section. It can then be overridden by other 'environment sections' but it must exist. As we use the same schema across all environments we only specify `dev:`.