Fix flagging licences for supplementary billing (#631)
https://eaflood.atlassian.net/browse/WATER-3995

During our SROC supplementary billing UAT it has been spotted that when the import runs and chooses to flag a licence for supplementary billing it is ignoring our new `include_in_sroc_supplementary_billing` flag. Understandable, we never knew the import service was doing anything with it!

So, we've checked what's going on and found this

```javascript
// src/modules/licence-import/load/licence.js

/*
  *  If the expired_date, lapsed_date or revoked_date are changed,
  *  flag the licence for supplementary billing
  */
if (licencePriorToImport && ((licencePriorToImport.expired_date !== licence.expiredDate) ||
    (licencePriorToImport.lapsed_date !== licence.lapsedDate) ||
    (licencePriorToImport.revoked_date !== licence.revokedDate)
)) {
  tasks.push(connectors.flagLicenceForSupplementaryBilling(licencePriorToImport.licence_id))
}
````

A quick and dirty solution would be to just add our field to that query and flag both. But a whole bunch of other things came out of UAT to do with these flags

- flagging licences for SROC when they have no SROC charge versions
- flagging licences for PRESROC when they have no PRESROC charge versions
- flagging licences with no charge versions at all!

In all cases, once flagged the only process for clearing it and the message that appears is for it to be processed in a supplementary bill run. But they never will be because they don't have the right charge versions or any at all.

So, with this change, we are going to fix all these things. We'll update the logic so that

- if a licence has both PRESROC and SROC charge versions it sets both flags
- if a licence only has one type of charge version it sets the appropriate flag
- if a licence has no charge versions it does nothing