Update supp. billing flag query to use start date
https://eaflood.atlassian.net/browse/WATER-3995

We've spotted when testing that some licences have charge versions with start dates >= 2022-04-01 yet their scheme is set to 'alcs'. This shouldn't be and we suspect another faulty import process is the cause.

Regardless, we need to update the query to use the `start_date` instead of `scheme` to determine if a charge version is SROC or not in order for the flagging to be correct.

🤦😩