Tidy up health routes and controller
The first thing we want to do is add support for Airbrake directly into project. When we do that in other projects we add an endpoint that allows us to test the integration is working. It lives on the `/health` route so we'll do the same here.

The key difference is our standard convention is to group routes and controllers together. This means `/health/info` and `/health/airbrake` would have their own controllers. In this repo things are grouped by module, so there is only one `controller.js` for `/health`.

That means we'll be adding our new endpoint to it. Only it's using the `const thing = () => {}` function convention which we don't like. Plus the current routes file doesn't match either our own or the project's convention.

So, this is an initial tidy to make our subsequent changes a easier.