Add GlobalNotifier to the project
This is our solution to logging also used by the [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system). Anything we log will use Hapi pino under the hood, so will be done in the same format as what it is automatically logging. It also gives us access to both log error and send a notification to Airbrake at the same time.