Replace legacy logger in licence import
This is the first import that uses onComplete() handlers and involves a chain of jobs. Rather than logging 'start' and 'finish' in both the job and onComplete() handlers we split them; the job handler logs the start and the onComplete() logs the finish.

It also involves creating a job for each licence and company to be imported. These are in the thousands. A primary driver for this change was to reduce noise. So, we've removed the logging from the individual jobs and now just log errors.

Other than that very little was consistent or tidy. So, we've done some clean up as we've touched each file.