Clean up params to jobs in licence-import
Only spotted because NALD import was next and it's plugin is a mess! But we were passing through args to onComplete() handlers that were not being used.

It's much clearer for those maintaining the code in future to be explicit about whether an arg is needed. So, took the time to clean this up.