Replace legacy logger in NALD import
This was fun!

This is the most complex module we've come across so far in water-abstraction-import. For example, most stuff (most!) was not using the logger directly but instead using a library module specific to NALD import. This wrapped the logger and made available a series of methods, all of which just stuck a 'helpful' message at the start of the log message.

That would need to go if NALD import was to match the changes we've done in the other modules.

We also found that the jobs rely on a number of services, a number of which will catch an error, log it, then just throw the error anyway! In most cases that error was being caught at the top level anyway, so catching it lower down was pointless.

In the others we've traced through that throwing rather than swallowing the error will have no impact on the process; the error was just getting caught and raised elsewhere.

The one exception is `transform-crm.js`. It does send something back even if an error is thrown. But we are confident the licence wouldn't be imported if errors were found so for that we have taken a calculated risk and removed the error handling. We'll test and monitor the logs to be sure that was the right decision.

Other than that the other big clean up was in the plugin. Someone hadn't gotten clever and chosen to generalise the PGBoss `subscribe()` and `onComplete()` calls. It hand't really resulted in any less code, especially when you take into account `jobs/index.js`. What it had done was make the code harder to follow, be inconsistent with the other plugins, and again result in args being passed that are not used.

We couldn't let this slide so whilst making our other changes, we got on and cleaned up the plugin. The general methods are gone and now its more consistent with the other modules. Plus we find it a lot easier to follow what is going on.