Replace legacy logging (#654)
https://eaflood.atlassian.net/browse/WATER-4039

The import service is a funny thing. It seems to throw numerous errors out with each run yet when someone takes notice it suddenly becomes important to understand why. Our problem is the existing logging

- it is extremely noisy; lots of messages and errors get logged
- it is lacking in any details or consistency; we can never figure out the source of the error nor which job caused it

We're not even sure when a job has started or finished.

This change aims to rectify that. We're bringing in our logging support in [water-abstraction-system](https://github.com/Defra/water-abstraction-system) into this project.

Airbrake will be directly integrated, using the latest version and automatically called when any unhandled exceptions are raised. We'll also bring in our `BaseNotifier` and `GlobalNotifier` modules, which along with our plugin, makes hapi-pino type logging available.

With those in place, we work through the various jobs and strip out all usage of the legacy logger. We log when a job starts and when it completes but nothing else. This removes the noise and helps us see when jobs are running.

Having to touch so many files meant we couldn't help ourselves. So, included with these changes is some housekeeping.

The legacy logger still remains, because it is required when instantiating the db connection using [water-abstraction-helpers](https://github.com/DEFRA/water-abstraction-helpers). So, we'll still have the `Pool low on connections::Total:20,Idle:0,Waiting:19` type messages appearing numerous times during a run. But we hope to tackle that on a subsequent commit.

**Notes**

- Tidy up health routes and controller

The first thing we want to do was add support for Airbrake directly into the project. When we do that in other projects we add an endpoint that allows us to test the integration is working. It lives on the `/health` route so we do the same here.

The key difference is our standard convention is to group routes and controllers together. This means `/health/info` and `/health/airbrake` would have their own controllers. In this repo things are grouped by module, so there is only one `controller.js` for `/health`.

That means we'll be adding our new endpoint to it. Only it's using the `const thing = () => {}` function convention which we don't like. Plus the current routes file doesn't match either our own or the project's convention.

- Replace legacy logger in index.js

Our experience on water-abstraction-system has shown we can access the logger from the server object when handling the `SIG` close events. Should things go pop though we are safer just using console.log(). Whilst we are at it we remove logging that was just noise and our testing found there was no need to do anything special with the DB pool.

- Remove unused file

We came across `src/modules/charging-import/lib/import-charge-version-metadata.js` when replacing the legacy logger calls and quickly found the module is not referenced anywhere else in the code.

- Logging in onComplete handlers

Some imports use `onComplete()` handlers and involve a chain of jobs. Rather than logging 'start' and 'finish' in both the job and onComplete() handlers we split them; the job handler logs the start and the onComplete() logs the finish.

- Logging in minor jobs

Some jobs are responsible for just importing a single licence or company. Other jobs are responsible for publishing these minor jobs to the queue. Most of the noise was coming from these jobs, which number in the thousands and were logging start, finish and stuff in between. We've chosen to only log when an error occurs as that is the only time we care about what is going on.

- Clean up params to jobs

We noticed that the handlers through the imports were always expecting a job and a message queue (in different orders depending on the module!) even though not all used them. As we control what gets passed, we spent some time cleaning up the params. It's much clearer for those maintaining the code in future to be explicit about whether an arg is needed.

- Update NALD plugin

Someone had gotten clever and chosen to generalise the PGBoss `subscribe()` and `onComplete()` calls in the plugin. But this hadn't really resulted in any less code, especially when you take into account `jobs/index.js`. What it had done was make the code harder to follow, be inconsistent with the other plugins, and again result in args being passed that are not used.

We couldn't let this slide so whilst making our other changes, we got on and cleaned up the plugin. The general methods are gone and now its more consistent with the other modules. Plus we could drop `index.js` in the jobs folder.

- Logging in nald-import other modules

This is the most complex module we've come across so far in water-abstraction-import. For example, most stuff (most!) was not using the logger directly but instead using a library module specific to NALD import. This wrapped the logger and made available a series of methods, all of which just stuck a 'helpful' message at the start of the log message.

We also found that the jobs rely on a number of services, of which some will catch an error, log it, and then re-throw the error! In most cases that error was being caught at the top level, so catching it lower down was pointless.

In the others we traced that throwing rather than swallowing the error will have no impact on the process; the error was just getting caught and raised elsewhere.

The one exception is `transform-crm.js`. It does send something back even if an error is thrown. But we are confident the licence wouldn't be imported if errors were found so for that we have taken a calculated risk and removed the error handling. We'll test and monitor the logs to be sure that was the right decision.