Refactor core plugin
Again with making everything generic! We did this not long ago in the NALD import plugin. We scrap all the usage of generic publish and subscribe methods and just get on with subscribing the jobs we have.

What makes it worse in this case is the fact _there is only one job!_ The generic methods were pointless. There was even logic to generate a timestamp which was being passed to the jobs createMessage() where it was utterly ignored.

Gah!

So, if only because a piece of us would have died we refactor the plugin to start making sense.