Remove block on import tracker in non-prod
We had no clue that an email is sent from the import service that reports on the success of the import jobs. A lot of this will be down to the fact there was a guard clause that only allowed it to be sent in production.

There is no reason why it can't be sent in non-production. We use a test Notify API key so no one is going to get an email. And this way we are bringing non-prod more inline with production.