Fix charging-import charge versions job
https://eaflood.atlassian.net/browse/WATER-4024

As part of building the new SROC supplementary billing engine flagging licences for supplementary billing suddenly became a hot topic.

We had to add a new flag to identify those licences flagged for SROC supplementary from those flagged for PRESROC. But what this highlighted was that there exists lots of legacy logic that has not been updated to recognise there are now 2 schemes.

For example, one of them was the the `licence-import`, which was [flagging everything for PRESROC](https://github.com/DEFRA/water-abstraction-import/pull/636) even if the licence start date was after 2022-04-01.

We spotted another one in the `src/modules/charging-import/jobs/charge-versions.js` job. The `importChargeVersions` query in `src/modules/charging-import/lib/queries/charging.js` sets the scheme to `alcs` (PRESROC) for records it creates.

The only 'funny' about this is the job and the query were built for a one off exercise! When billing switched to the WRLS service from NALD, charge versions needed to be created for every licence. This job did that but was only intended to be run as a one off.

However, when the current DEV team together we were informed we would need to trigger this job (done by a web request) to get our local environments setup!

So, because it is occasionally still used and if only to avoid forgetting all this and panicking again when we spot it in the future we are fixing it.

This change sets the scheme for the charge version based on the start date we are applying.