Update query for charge versions meta data import
Whilst sorting out the query used by the charge versions import we found another occurrence of scheme being hard coded as 'alcs'. It is in a query that appears to be used in the NALD import process.

Specifically, it's to update the `water_import.charge_versions_meta_data` table which is the basis for the charge versions import. It seems we are always running this query, processing each licence during the NALD import.

When we looked at the existing LIVE data no records in `water_import.charge_versions_meta_data` had a start date after 2022-04-01. But just in case, we apply the same intelligence we used in the `importChargeVersions` query.