Alter fix for charge versions import
https://eaflood.atlassian.net/browse/WATER-4024

We implemented [Fix charging-import charge versions job](https://github.com/DEFRA/water-abstraction-import/pull/658) because we had spotted that an existing import job (charge versions) did not have knowledge of SROC so was hard coding everything as `alcs`.

Having merged all pleased with ourselves, we then spotted the charge version import is followed by a charge element import. The same problem exists there, the data it is setting up and the child records it is creating are all for the PRESROC scheme. It would need a complete re-write to cater for the SROC scheme! 😱😩

So, a better fix would be to alter the charge version import to _only_ do it for PRESROC licences. Anything with a start date 2022-04-01 or greater will be ignored and folks setting up a new environment will just have to create SROC charge versions manually (until we have a proper dev/test data seeding solution).