Update getNonDraftChargeVersionsForLicence query
We learnt in the previous change that this query is used to populate the `water_import.charge_version_metadata` table and it's the content of that table which drives what charge versions get imported.

So, we revert our logic in how to set the scheme and add a new where clauses to only return NALD charge versions with a start date in the PRESROC period. This means the source data for the charge versions import will only be PRESROC based.