Add where clause to charge versions query
We know having checked locally there are records in the `water_import.charge_versions_metadata` tables with start dates after 2022-04-01. This means it is possible the job will run on a system that has data we now don't want in the table.

So, to avoid any issues and weirdness we put this WHERE clause in just in case.