Move ChargeVersionsMetadataImport out of NALD (#661)
https://eaflood.atlassian.net/browse/WATER-4024

This all originated when we spotted what looked like an automated mechanism for importing charge versions into the `water` schema as part of the NALD import job. Over a series of changes to the import service and discussions with the team, we learnt it's a very mixed picture.

Yes, the charge versions import did not understand there is an SROC scheme so could potentially be creating `alcs` charge versions with start dates after 2022-04-01 (the SROC start date). But the charge version import was a one-off job, used to prepare the service for taking over billing.

The `water_import.charge_versions_metadata` is populated as part of the NALD import process. But it is only used by `src/modules/charging-import/jobs/charge-versions.js` which we now know is redundant.

This means each day the NALD import is processing 70K licences through this step for no reason. To help improve the performance of the process we're moving this work to where it should have been all along, in the charging-import `charge-versions.js` job.

We are moving it rather than deleting it because the dev and test teams still rely on the charge versions import. When building or rebuilding a local environment we need this job to create the PRESROC charge versions after the licences have been imported.