Update log and errbit notifiers
When working on [Move ChargeVersionsMetadataImport out of NALD](https://github.com/DEFRA/water-abstraction-import/pull/661) we encountered an error. That was expected; we were mid-implementation. What we didn't expect was

- the lack of detail in the log
- that attempting to view the error in Errbit caused an error

Suffice to say, we realised our notifier implementation was wrong. We were passing things incorrectly to both [pino](https://github.com/pinojs/pino) and [errbit](https://github.com/errbit/errbit). [Better handle errors in Notifiers](https://github.com/DEFRA/water-abstraction-system/pull/273) goes into more detail about the issue and what we did to resolve it.

That is our main repo and the place we copied the current implementation of `GlobalNotifier` and `BaseNotifier` from. So, we did the original fixes there. This change updates the copies of those notifiers in this repo to include the fixes.