Add logging for company and licence import
https://eaflood.atlassian.net/browse/WATER-4039

After updating the logging in the app to better diagnose tickets like WATER-4039 we've been checking all is well.

It seems it is. We see no errors in the logs. But we are also not seeing the `date_updated` change in our testing environment for `water.licences`.

We removed the logging in import company and import licence, because these generated 1000's of log lines daily. But we need to know what's going on.

Fortunately, `licence-import/handlers/on-complete-import-company.js` already seems to be checking if there is anything pending. So, we can use this to add a log message that the job is complete (we should only see the one entry).

`licence-import/handlers/import-licence.js` on the overhand doesn't have this check. If we were on a later version of pgBoss we could just ask the queue for its pending job count. But we're not. So for now we've added the option to start logging each job again, though put it behind the debug log level. That way we can switch it on when we think we have a problem, and turn it off for the rest of the time.